requirements.txt
streamlit
pandas
PyPDF2
python-docx
rapidfuzz
sentence-transformers
torch
transformers
matplotlib
wordcloud
